+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactor NavMenu component to use context for inventory data and simplify props
 - Refactor MainContainer to use context for inventory management and update tests accordingly
+- Refactor QuickStatsBar to use context for inventory management
 
 ### Deprecated